accountant module working to be continue

+++ SmartPay_Admin/aspnet-core/src/SmartPay.EntityFrameworkCore/Migrations/SmartPayDbContextModelSnapshot.cs
@@ -1758,8 +1758,8 @@ namespace SmartPay.Migrations
                     b.Property<long?>("LastModifierUserId")
                         .HasColumnType("bigint");
 
-                    b.Property<int>("PhoneNumber")
-                        .HasColumnType("int");
+                    b.Property<long>("PhoneNumber")
+                        .HasColumnType("bigint");
 
                     b.Property<long>("UserId")
                         .HasColumnType("bigint");
